requirements.txt
pandas
numpy
scikit-learn
faker
matplotlib
seaborn
wordcloud
nltk
google-cloud-bigquery